git-svn-id: http://svn.resourcespace.com/svn/rs/branches/20230919_acota_q12903@22971 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/message_functions.php
@@ -424,9 +424,12 @@ function message_deleteselusrmsg($messages)
  
     $parameters = array("i",(int)$userref);
     $messages = json_decode($messages, true);
-    $parameters = array_merge($parameters, ps_param_fill($messages,"i"));
- 
-    ps_query("DELETE FROM user_message WHERE user = ? AND ref IN (" . ps_param_insert(count($messages)) . ")", $parameters);
+    if (count($messages) > 0) 
+        {
+        $parameters = array_merge($parameters, ps_param_fill($messages,"i"));
+
+        ps_query("DELETE FROM user_message WHERE user = ? AND ref IN (" . ps_param_insert(count($messages)) . ")", $parameters);
+        }
     }
  
 /**
